+++ src/graphql_clj/spec.clj
@@ -48,7 +48,7 @@
   [s path]
   (cond (default-type-names (first path)) (keyword base-ns (first path))
         (keyword? path)                   path
-        (vector? path)                    (keyword (spec-namespace s path) (name (last path)))
+        (or (vector? path) (seq? path))   (keyword (spec-namespace s path) (name (last path)))
         :else                             (ge/throw-error "Unhandled named-spec case" {:path path})))
 
 (defn- type-names->args [type-names]
@@ -118,6 +118,9 @@
 (defmethod spec-for :fragment-definition [s {:keys [v/path] :as n}]
   (register-idempotent (dissoc s :schema-hash) ["frag" (:name n)] (named-spec s path))) ;; alternatively, (to-keys s selection-set)
 
+(defmethod spec-for :inline-fragment [s {:keys [v/path] :as n}]
+  [(named-spec s [(last path)])])
+
 (defmethod spec-for :fragment-spread [s n]
   [(named-spec (dissoc s :schema-hash) ["frag" (:name n)])])
 
@@ -146,8 +149,10 @@
 (defmethod spec-for :input-type-field [s n]
   (register-type-field s n))
 
-(defmethod spec-for :field [s {:keys [v/path]}]
-  [(named-spec s path)])
+(defmethod spec-for :field [s {:keys [v/path v/parent]}]
+  [(named-spec s (if (= :inline-fragment (:node-type parent))
+                   [(last (butlast path)) (last path)] ;; Ignore hierarchy for inline fragments
+                   path))])
 
 (defmethod spec-for :argument [s {:keys [v/path]}]
   [(named-spec s (into ["arg"] path))])
@@ -156,8 +161,6 @@
   (register-idempotent s (into ["arg"] path) (named-spec s [(to-type-name n)])))
 
 (defmethod spec-for :default [_ _])
-
-;; Parent and base types
 
 (defmulti ^:private of-type (fn [n _] (:node-type n)))
 
@@ -170,29 +173,41 @@
 (defmethod ^:private of-type :default [{:keys [spec]} _]
   spec)
 
-(defn get-type-node
+;; Parent and base types
+
+(defn get-type-node [spec s]
+  "Given a spec, get the corresponding node from the AST"
+  (get-in s [:spec-map spec]))
+
+(defn get-base-type-node
   "Given a spec, get the node definition for the corresponding base type"
   [{:keys [spec]} s]
-  (let [base-spec (s/get-spec spec)]
+  (when-let [base-spec (s/get-spec spec)]
     (if (default-type-names (name base-spec))
       {:node-type :scalar :type-name (name base-spec)}
-      (get-in s [:spec-map base-spec]))))
+      (get-type-node base-spec s))))
 
 (defn get-parent-type
   "Given a node and the global state, find the parent type"
   [{:keys [v/parent]} s]
-  (if-let [base-parent (get-in s [:spec-map (of-type parent s)])]
+  (if-let [base-parent (get-type-node (of-type parent s) s)]
     (of-type base-parent s)
     (recur parent s)))
 
 ;; Visitors
 
+(defn- safe-eval [d]
+  (assert (= (first d) 'clojure.spec/def))               ;; Protect against unexpected statement eval
+  (try (eval d) (catch Compiler$CompilerException _ d))) ;; Squashing errors here to provide better error messages in validation
+
 (def define-specs)
 (zv/defvisitor define-specs :post [n s]
-  (when (seq? n)
-    (doseq [d (some-> s :spec-defs)]
-      (assert (= (first d) 'clojure.spec/def))              ;; Protect against unexpected statement eval
-      (try (eval d) (catch Compiler$CompilerException _)))  ;; Squashing errors here to provide better error messages in validation
+  (when (seq? n) ;; Top of the tree is a seq
+    (some->>
+      (some-> s :spec-defs)
+      (mapv safe-eval)
+      (filter (comp not keyword?))
+      (mapv safe-eval)) ;; if eval failed the first time, try once more to help with order dependencies
     {:state (dissoc s :spec-defs)}))
 
 (def keywordize)