+++ src/graphql_clj/validator/rules/variables_are_input_types.clj
@@ -10,7 +10,7 @@
 (def acceptable-types #{:scalar :enum-definition :input-definition})
 
 (defnodevisitor bad-variable-type :pre :variable-definition [n s]
-  (let [{:keys [node-type] :as type-node} (spec/get-type-node n s)]
+  (let [{:keys [node-type] :as type-node} (spec/get-base-type-node n s)]
     (when-not (acceptable-types node-type)
       {:state (ve/update-errors s (bad-variable-type-error n type-node))})))