+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased][unreleased]
 - Validate that statements do not contain fragment cycles
+- Validate that fragments are on composite types
+- Make validation robust to order (can now reference types declared later in a statement or schema)
 
 ## [0.1.17] - 2016-10-14
 - Implement preliminary version of FieldsOnCorrectType validation